🐛 demolish useUserInfo, unify userdata management
