+++ composables/useLogin.ts
@@ -2,6 +2,16 @@
 import useApiClient from '~/composables/useApiClient'
 import { useAuthStore, User } from '~/stores/auth'
 
+export type UserInfo = {
+  name_first: string
+  name_first_kana: string
+  name_last: string
+  name_last_kana: string
+  email: string
+  gender: number
+  student_id: string
+}
+
 export const useLogin = () => {
   const authStore = useAuthStore()
   const authWithCode = async (code: string): Promise<boolean> => {
@@ -60,5 +70,29 @@ export const useLogin = () => {
     authStore.clearToken()
     authStore.clearUser()
   }
-  return { getCurrentUser, signOut, trySignIn, authWithCode }
+
+  // this is client for using credentials
+  const client = useApiClient()
+
+  const updateUserInfo = async (ui: UserInfo): Promise<boolean> => {
+    try {
+      const resp = await client.post('/api/v1/user/info', ui)
+      if (resp.error) {
+        return false
+      }
+      authStore.clearUser()
+      return true
+    } catch (err) {
+      console.error(err)
+      return false
+    }
+  }
+
+  return {
+    getCurrentUser,
+    signOut,
+    trySignIn,
+    authWithCode,
+    updateUserInfo
+  }
 }
